requirements.txt
einops 
datasets 
matplotlib 
tqdm 
accelerate 
jax[cpu] 
torchinfo 
wandb
ema_pytorch
lpips
pyyaml
diffusers
transformers
torch-ema